use column id in logical scan

+++ rust/frontend/src/catalog/mod.rs
@@ -9,10 +9,10 @@ mod database_catalog;
 mod schema_catalog;
 mod table_catalog;
 
-pub(crate) type DatabaseId = u32;
-pub(crate) type SchemaId = u32;
-pub(crate) type TableId = u32;
-pub(crate) type ColumnId = u32;
+pub(crate) type DatabaseId = u64;
+pub(crate) type SchemaId = u64;
+pub(crate) type TableId = u64;
+pub(crate) type ColumnId = u64;
 
 #[derive(Error, Debug)]
 pub enum CatalogError {
